refactor(swap): extract quote/fee orchestration into SwapQuotePipelineController (#4865)
Move the gas/network-fee state and quote pipeline wiring (calculateGas,
collectTotalFee, observeQuotePipeline, applyQuoteResult,
applyNetworkFeeOutcome, showIndicativeRate, launchRefreshQuoteTimer,
updatePairSupport, resetQuoteState) out of SwapFormViewModel into a
dedicated SwapQuotePipelineController collaborator. The controller owns
the fee/quote StateFlows and writes results back to the shared uiState;
the ViewModel only constructs it, starts it, and reads the resolved
quote/fee values for swap(), flip, and percentage taps.

VM drops from 858 to 459 lines.
